Add API file and  go.sum
In Postman, you can export a collection of API requests as an `api.json` file. This file contains the details of all the requests in the collection, including the request URL, HTTP method, headers, and body. It also includes information about the expected response, such as the response status code, headers, and body schema.

`go.sum`: This file is used by Go to ensure that the correct versions of dependencies are used in the project. It contains a list of all the modules used in the project, along with their versions and checksums. When the project is built, Go checks the `go.sum` file to ensure that the correct versions of the dependencies are downloaded and used. If a dependency's version or checksum has changed since the last build, Go will download the new version and update the `go.sum` file. This helps ensure that the project is built with consistent and secure dependencies.