Add demo file for others to use as a reference when writing code
GitHub Copilot: The Git commit message "Add demo file for others to use as a reference when writing code" indicates that the author has added a demo file to the project that can be used as a reference when writing code. The demo file is likely to contain example code or templates that demonstrate how to use certain features or implement certain functionality in the project. The purpose of adding this file is to make it easier for other developers to understand how to use the project and to reduce the amount of time they need to spend figuring out how to write code that works with the project. This commit message is clear and informative, and provides a useful summary of the changes made in the commit.