Add Person, Account, Persons, and Accounts messages to model.proto
This commit adds the `Person`, `Account`, `Persons`, and `Accounts` messages to the `model.proto` file. The `Person` message defines a person object with various fields, while the `Account` message defines an account object that includes a `Person` object as well as additional fields. The `Persons` and `Accounts` messages define collections of `Person` and `Account` objects, respectively. This update will allow for the serialization and deserialization of these objects using Protocol Buffers.