Create go.yml
This GitHub Actions workflow is designed to automate building and testing a Golang project. The main job is named "build," which performs the following steps:

Checks whether Protocol Buffers compiler (protoc) is already installed. If not, downloads and installs it from the official repository.
#    Sets up Go by using the actions/setup-go action, specifying version 1.21 as the required version.
#    Runs go mod tidy to ensure all necessary packages are included in the module's dependency graph.
#    Runs go mod vendor to create a copy of the current state of the module's dependencies in the vendor directory.
#    Executes go build -v ./..., which builds the entire package and its dependencies.
#    Finally, executes go test -v ./..., which runs tests for the entire package and its dependencies.

The workflow triggers on both pushes and pull requests to the main branch. Additionally, it specifies Ubuntu as the operating system for running the jobs.